Changes for the 9 years including:
. node41 plugin;
. meta-data and inline-data checksums support;
. discard support for SSD devices;
. many bugfixes.

+++ demos/trunc.c
@@ -37,7 +37,7 @@ errno_t trunc_cmd(busy_ctx_t *ctx) {
 	/* Truncate the object. */
 	if (reiser4_object_truncate(object, ctx->count)) {
 		aal_error("Can't truncate object %s to %lld bytes.", 
-			  ctx->in.path, ctx->count);
+			  ctx->in.path, (long long)ctx->count);
 		goto error_close_object;
 	}
